libs/xpcom: Shut up some fprintf warnings.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@109062 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/libs/xpcom18a4/ipc/ipcd/client/src/ipcDConnectService.cpp
@@ -2996,11 +2996,11 @@ ipcDConnectService::Shutdown()
           " => requests processed:        %llu\n"
           " => avg requests process time: %llu ns\n"
           " => avg requests waiting time: %llu ns\n",
-          RTReqPoolGetStat(mhReqPool, RTREQPOOLSTAT_THREADS),
-          RTReqPoolGetStat(mhReqPool, RTREQPOOLSTAT_THREADS_CREATED),
-          RTReqPoolGetStat(mhReqPool, RTREQPOOLSTAT_REQUESTS_PROCESSED),
-          RTReqPoolGetStat(mhReqPool, RTREQPOOLSTAT_NS_AVERAGE_REQ_PROCESSING),
-          RTReqPoolGetStat(mhReqPool, RTREQPOOLSTAT_NS_AVERAGE_REQ_QUEUED)
+          (unsigned long long)RTReqPoolGetStat(mhReqPool, RTREQPOOLSTAT_THREADS),
+          (unsigned long long)RTReqPoolGetStat(mhReqPool, RTREQPOOLSTAT_THREADS_CREATED),
+          (unsigned long long)RTReqPoolGetStat(mhReqPool, RTREQPOOLSTAT_REQUESTS_PROCESSED),
+          (unsigned long long)RTReqPoolGetStat(mhReqPool, RTREQPOOLSTAT_NS_AVERAGE_REQ_PROCESSING),
+          (unsigned long long)RTReqPoolGetStat(mhReqPool, RTREQPOOLSTAT_NS_AVERAGE_REQ_QUEUED)
           );
 #endif
